make LoggingUtils configurable
