+++ core/cas-server-core-util-api/src/main/java/org/apereo/cas/util/LoggingUtils.java
@@ -9,9 +9,9 @@ import org.apache.commons.lang3.exception.ExceptionUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.Arrays;
 import java.util.Map;
 import java.util.Objects;
+import java.util.ServiceLoader;
 
 /**
  * This is {@link LoggingUtils}.
@@ -22,13 +22,15 @@ import java.util.Objects;
 @UtilityClass
 public class LoggingUtils {
 
-    private static boolean SUMMARIZED_STACK_TRACES_DISABLED;
+    private static final LogMessageSummarizer LOG_MESSAGE_SUMMARIZER;
 
     /*
      * Allow customization of whether this class will summarize stack traces when log level higher than debug.
      */
     static {
-        SUMMARIZED_STACK_TRACES_DISABLED = Boolean.getBoolean("cas.summarizedStackTracesDisabled");
+        LOG_MESSAGE_SUMMARIZER = ServiceLoader.load(LogMessageSummarizer.class)
+                .findFirst()
+                .orElse(new DefaultLogMessageSummarizer());
     }
 
     private static final int CHAR_REPEAT_ACCOUNT = 60;
@@ -92,9 +94,9 @@ public class LoggingUtils {
      * @param throwable the throwable
      */
     public static void error(final Logger logger, final String msg, final Throwable throwable) {
-        FunctionUtils.doIf(SUMMARIZED_STACK_TRACES_DISABLED || logger.isDebugEnabled(),
+        FunctionUtils.doIf(LOG_MESSAGE_SUMMARIZER.shouldSummarize(logger),
                 unused -> logger.error(msg, throwable),
-                unused -> logger.error(summarizeStackTrace(msg, throwable)))
+                unused -> logger.error(LOG_MESSAGE_SUMMARIZER.summarizeStackTrace(msg, throwable)))
             .accept(throwable);
     }
 
@@ -126,19 +128,10 @@ public class LoggingUtils {
      * @param throwable the throwable
      */
     public static void warn(final Logger logger, final String message, final Throwable throwable) {
-        FunctionUtils.doIf(SUMMARIZED_STACK_TRACES_DISABLED || logger.isDebugEnabled(),
+        FunctionUtils.doIf(LOG_MESSAGE_SUMMARIZER.shouldSummarize(logger),
                 unused -> logger.warn(message, throwable),
-                unused -> logger.warn(summarizeStackTrace(message, throwable)))
+                unused -> logger.warn(LOG_MESSAGE_SUMMARIZER.summarizeStackTrace(message, throwable)))
             .accept(throwable);
-    }
-
-    private static String summarizeStackTrace(final String message, final Throwable throwable) {
-        val builder = new StringBuilder(message).append('\n');
-        Arrays.stream(throwable.getStackTrace()).limit(3).forEach(trace -> {
-            val error = String.format("\t%s:%s:%s%n", trace.getFileName(), trace.getMethodName(), trace.getLineNumber());
-            builder.append(error);
-        });
-        return builder.toString();
     }
 
     /**
