+++ core/cas-server-core-util-api/src/main/resources/services/org.apereo.cas.util.LogMessageSummarizer
@@ -1,0 +1,1 @@
+org.apereo.cas.util.DefaultLogMessageSummarizer
